Add files via upload
This game is written with python's pygame library and made use of modules and etc. to be closer to real life applications and decency of professionality. Another important is that, this game is an appearing of curiosity and enthusiasm of self improvement. It lacks of concepts, codes, etc and it needs to get bug free.